Move some functions from issue.go to standalone files
