+++ routers/web/repo/issue.go
@@ -39,7 +39,6 @@ import (
 	"code.gitea.io/gitea/modules/markup"
 	"code.gitea.io/gitea/modules/markup/markdown"
 	"code.gitea.io/gitea/modules/optional"
-	repo_module "code.gitea.io/gitea/modules/repository"
 	"code.gitea.io/gitea/modules/setting"
 	api "code.gitea.io/gitea/modules/structs"
 	"code.gitea.io/gitea/modules/templates"
@@ -2484,852 +2483,6 @@ func UpdateIssueAssignee(ctx *context.Context) {
 	ctx.JSONOK()
 }
 
-// UpdatePullReviewRequest add or remove review request
-func UpdatePullReviewRequest(ctx *context.Context) {
-	issues := getActionIssues(ctx)
-	if ctx.Written() {
-		return
-	}
-
-	reviewID := ctx.FormInt64("id")
-	action := ctx.FormString("action")
-
-	// TODO: Not support 'clear' now
-	if action != "attach" && action != "detach" {
-		ctx.Status(http.StatusForbidden)
-		return
-	}
-
-	for _, issue := range issues {
-		if err := issue.LoadRepo(ctx); err != nil {
-			ctx.ServerError("issue.LoadRepo", err)
-			return
-		}
-
-		if !issue.IsPull {
-			log.Warn(
-				"UpdatePullReviewRequest: refusing to add review request for non-PR issue %-v#%d",
-				issue.Repo, issue.Index,
-			)
-			ctx.Status(http.StatusForbidden)
-			return
-		}
-		if reviewID < 0 {
-			// negative reviewIDs represent team requests
-			if err := issue.Repo.LoadOwner(ctx); err != nil {
-				ctx.ServerError("issue.Repo.LoadOwner", err)
-				return
-			}
-
-			if !issue.Repo.Owner.IsOrganization() {
-				log.Warn(
-					"UpdatePullReviewRequest: refusing to add team review request for %s#%d owned by non organization UID[%d]",
-					issue.Repo.FullName(), issue.Index, issue.Repo.ID,
-				)
-				ctx.Status(http.StatusForbidden)
-				return
-			}
-
-			team, err := organization.GetTeamByID(ctx, -reviewID)
-			if err != nil {
-				ctx.ServerError("GetTeamByID", err)
-				return
-			}
-
-			if team.OrgID != issue.Repo.OwnerID {
-				log.Warn(
-					"UpdatePullReviewRequest: refusing to add team review request for UID[%d] team %s to %s#%d owned by UID[%d]",
-					team.OrgID, team.Name, issue.Repo.FullName(), issue.Index, issue.Repo.ID)
-				ctx.Status(http.StatusForbidden)
-				return
-			}
-
-			_, err = issue_service.TeamReviewRequest(ctx, issue, ctx.Doer, team, action == "attach")
-			if err != nil {
-				if issues_model.IsErrNotValidReviewRequest(err) {
-					log.Warn(
-						"UpdatePullReviewRequest: refusing to add invalid team review request for UID[%d] team %s to %s#%d owned by UID[%d]: Error: %v",
-						team.OrgID, team.Name, issue.Repo.FullName(), issue.Index, issue.Repo.ID,
-						err,
-					)
-					ctx.Status(http.StatusForbidden)
-					return
-				}
-				ctx.ServerError("TeamReviewRequest", err)
-				return
-			}
-			continue
-		}
-
-		reviewer, err := user_model.GetUserByID(ctx, reviewID)
-		if err != nil {
-			if user_model.IsErrUserNotExist(err) {
-				log.Warn(
-					"UpdatePullReviewRequest: requested reviewer [%d] for %-v to %-v#%d is not exist: Error: %v",
-					reviewID, issue.Repo, issue.Index,
-					err,
-				)
-				ctx.Status(http.StatusForbidden)
-				return
-			}
-			ctx.ServerError("GetUserByID", err)
-			return
-		}
-
-		_, err = issue_service.ReviewRequest(ctx, issue, ctx.Doer, &ctx.Repo.Permission, reviewer, action == "attach")
-		if err != nil {
-			if issues_model.IsErrNotValidReviewRequest(err) {
-				log.Warn(
-					"UpdatePullReviewRequest: refusing to add invalid review request for %-v to %-v#%d: Error: %v",
-					reviewer, issue.Repo, issue.Index,
-					err,
-				)
-				ctx.Status(http.StatusForbidden)
-				return
-			}
-			if issues_model.IsErrReviewRequestOnClosedPR(err) {
-				ctx.Status(http.StatusForbidden)
-				return
-			}
-			ctx.ServerError("ReviewRequest", err)
-			return
-		}
-	}
-
-	ctx.JSONOK()
-}
-
-// SearchIssues searches for issues across the repositories that the user has access to
-func SearchIssues(ctx *context.Context) {
-	before, since, err := context.GetQueryBeforeSince(ctx.Base)
-	if err != nil {
-		ctx.Error(http.StatusUnprocessableEntity, err.Error())
-		return
-	}
-
-	var isClosed optional.Option[bool]
-	switch ctx.FormString("state") {
-	case "closed":
-		isClosed = optional.Some(true)
-	case "all":
-		isClosed = optional.None[bool]()
-	default:
-		isClosed = optional.Some(false)
-	}
-
-	var (
-		repoIDs   []int64
-		allPublic bool
-	)
-	{
-		// find repos user can access (for issue search)
-		opts := &repo_model.SearchRepoOptions{
-			Private:     false,
-			AllPublic:   true,
-			TopicOnly:   false,
-			Collaborate: optional.None[bool](),
-			// This needs to be a column that is not nil in fixtures or
-			// MySQL will return different results when sorting by null in some cases
-			OrderBy: db.SearchOrderByAlphabetically,
-			Actor:   ctx.Doer,
-		}
-		if ctx.IsSigned {
-			opts.Private = true
-			opts.AllLimited = true
-		}
-		if ctx.FormString("owner") != "" {
-			owner, err := user_model.GetUserByName(ctx, ctx.FormString("owner"))
-			if err != nil {
-				if user_model.IsErrUserNotExist(err) {
-					ctx.Error(http.StatusBadRequest, "Owner not found", err.Error())
-				} else {
-					ctx.Error(http.StatusInternalServerError, "GetUserByName", err.Error())
-				}
-				return
-			}
-			opts.OwnerID = owner.ID
-			opts.AllLimited = false
-			opts.AllPublic = false
-			opts.Collaborate = optional.Some(false)
-		}
-		if ctx.FormString("team") != "" {
-			if ctx.FormString("owner") == "" {
-				ctx.Error(http.StatusBadRequest, "", "Owner organisation is required for filtering on team")
-				return
-			}
-			team, err := organization.GetTeam(ctx, opts.OwnerID, ctx.FormString("team"))
-			if err != nil {
-				if organization.IsErrTeamNotExist(err) {
-					ctx.Error(http.StatusBadRequest, "Team not found", err.Error())
-				} else {
-					ctx.Error(http.StatusInternalServerError, "GetUserByName", err.Error())
-				}
-				return
-			}
-			opts.TeamID = team.ID
-		}
-
-		if opts.AllPublic {
-			allPublic = true
-			opts.AllPublic = false // set it false to avoid returning too many repos, we could filter by indexer
-		}
-		repoIDs, _, err = repo_model.SearchRepositoryIDs(ctx, opts)
-		if err != nil {
-			ctx.Error(http.StatusInternalServerError, "SearchRepositoryIDs", err.Error())
-			return
-		}
-		if len(repoIDs) == 0 {
-			// no repos found, don't let the indexer return all repos
-			repoIDs = []int64{0}
-		}
-	}
-
-	keyword := ctx.FormTrim("q")
-	if strings.IndexByte(keyword, 0) >= 0 {
-		keyword = ""
-	}
-
-	isPull := optional.None[bool]()
-	switch ctx.FormString("type") {
-	case "pulls":
-		isPull = optional.Some(true)
-	case "issues":
-		isPull = optional.Some(false)
-	}
-
-	var includedAnyLabels []int64
-	{
-		labels := ctx.FormTrim("labels")
-		var includedLabelNames []string
-		if len(labels) > 0 {
-			includedLabelNames = strings.Split(labels, ",")
-		}
-		includedAnyLabels, err = issues_model.GetLabelIDsByNames(ctx, includedLabelNames)
-		if err != nil {
-			ctx.Error(http.StatusInternalServerError, "GetLabelIDsByNames", err.Error())
-			return
-		}
-	}
-
-	var includedMilestones []int64
-	{
-		milestones := ctx.FormTrim("milestones")
-		var includedMilestoneNames []string
-		if len(milestones) > 0 {
-			includedMilestoneNames = strings.Split(milestones, ",")
-		}
-		includedMilestones, err = issues_model.GetMilestoneIDsByNames(ctx, includedMilestoneNames)
-		if err != nil {
-			ctx.Error(http.StatusInternalServerError, "GetMilestoneIDsByNames", err.Error())
-			return
-		}
-	}
-
-	projectID := optional.None[int64]()
-	if v := ctx.FormInt64("project"); v > 0 {
-		projectID = optional.Some(v)
-	}
-
-	// this api is also used in UI,
-	// so the default limit is set to fit UI needs
-	limit := ctx.FormInt("limit")
-	if limit == 0 {
-		limit = setting.UI.IssuePagingNum
-	} else if limit > setting.API.MaxResponseItems {
-		limit = setting.API.MaxResponseItems
-	}
-
-	searchOpt := &issue_indexer.SearchOptions{
-		Paginator: &db.ListOptions{
-			Page:     ctx.FormInt("page"),
-			PageSize: limit,
-		},
-		Keyword:             keyword,
-		RepoIDs:             repoIDs,
-		AllPublic:           allPublic,
-		IsPull:              isPull,
-		IsClosed:            isClosed,
-		IncludedAnyLabelIDs: includedAnyLabels,
-		MilestoneIDs:        includedMilestones,
-		ProjectID:           projectID,
-		SortBy:              issue_indexer.SortByCreatedDesc,
-	}
-
-	if since != 0 {
-		searchOpt.UpdatedAfterUnix = optional.Some(since)
-	}
-	if before != 0 {
-		searchOpt.UpdatedBeforeUnix = optional.Some(before)
-	}
-
-	if ctx.IsSigned {
-		ctxUserID := ctx.Doer.ID
-		if ctx.FormBool("created") {
-			searchOpt.PosterID = optional.Some(ctxUserID)
-		}
-		if ctx.FormBool("assigned") {
-			searchOpt.AssigneeID = optional.Some(ctxUserID)
-		}
-		if ctx.FormBool("mentioned") {
-			searchOpt.MentionID = optional.Some(ctxUserID)
-		}
-		if ctx.FormBool("review_requested") {
-			searchOpt.ReviewRequestedID = optional.Some(ctxUserID)
-		}
-		if ctx.FormBool("reviewed") {
-			searchOpt.ReviewedID = optional.Some(ctxUserID)
-		}
-	}
-
-	// FIXME: It's unsupported to sort by priority repo when searching by indexer,
-	//        it's indeed an regression, but I think it is worth to support filtering by indexer first.
-	_ = ctx.FormInt64("priority_repo_id")
-
-	ids, total, err := issue_indexer.SearchIssues(ctx, searchOpt)
-	if err != nil {
-		ctx.Error(http.StatusInternalServerError, "SearchIssues", err.Error())
-		return
-	}
-	issues, err := issues_model.GetIssuesByIDs(ctx, ids, true)
-	if err != nil {
-		ctx.Error(http.StatusInternalServerError, "FindIssuesByIDs", err.Error())
-		return
-	}
-
-	ctx.SetTotalCountHeader(total)
-	ctx.JSON(http.StatusOK, convert.ToIssueList(ctx, ctx.Doer, issues))
-}
-
-func getUserIDForFilter(ctx *context.Context, queryName string) int64 {
-	userName := ctx.FormString(queryName)
-	if len(userName) == 0 {
-		return 0
-	}
-
-	user, err := user_model.GetUserByName(ctx, userName)
-	if user_model.IsErrUserNotExist(err) {
-		ctx.NotFound("", err)
-		return 0
-	}
-
-	if err != nil {
-		ctx.Error(http.StatusInternalServerError, err.Error())
-		return 0
-	}
-
-	return user.ID
-}
-
-// ListIssues list the issues of a repository
-func ListIssues(ctx *context.Context) {
-	before, since, err := context.GetQueryBeforeSince(ctx.Base)
-	if err != nil {
-		ctx.Error(http.StatusUnprocessableEntity, err.Error())
-		return
-	}
-
-	var isClosed optional.Option[bool]
-	switch ctx.FormString("state") {
-	case "closed":
-		isClosed = optional.Some(true)
-	case "all":
-		isClosed = optional.None[bool]()
-	default:
-		isClosed = optional.Some(false)
-	}
-
-	keyword := ctx.FormTrim("q")
-	if strings.IndexByte(keyword, 0) >= 0 {
-		keyword = ""
-	}
-
-	var labelIDs []int64
-	if splitted := strings.Split(ctx.FormString("labels"), ","); len(splitted) > 0 {
-		labelIDs, err = issues_model.GetLabelIDsInRepoByNames(ctx, ctx.Repo.Repository.ID, splitted)
-		if err != nil {
-			ctx.Error(http.StatusInternalServerError, err.Error())
-			return
-		}
-	}
-
-	var mileIDs []int64
-	if part := strings.Split(ctx.FormString("milestones"), ","); len(part) > 0 {
-		for i := range part {
-			// uses names and fall back to ids
-			// non existent milestones are discarded
-			mile, err := issues_model.GetMilestoneByRepoIDANDName(ctx, ctx.Repo.Repository.ID, part[i])
-			if err == nil {
-				mileIDs = append(mileIDs, mile.ID)
-				continue
-			}
-			if !issues_model.IsErrMilestoneNotExist(err) {
-				ctx.Error(http.StatusInternalServerError, err.Error())
-				return
-			}
-			id, err := strconv.ParseInt(part[i], 10, 64)
-			if err != nil {
-				continue
-			}
-			mile, err = issues_model.GetMilestoneByRepoID(ctx, ctx.Repo.Repository.ID, id)
-			if err == nil {
-				mileIDs = append(mileIDs, mile.ID)
-				continue
-			}
-			if issues_model.IsErrMilestoneNotExist(err) {
-				continue
-			}
-			ctx.Error(http.StatusInternalServerError, err.Error())
-		}
-	}
-
-	projectID := optional.None[int64]()
-	if v := ctx.FormInt64("project"); v > 0 {
-		projectID = optional.Some(v)
-	}
-
-	isPull := optional.None[bool]()
-	switch ctx.FormString("type") {
-	case "pulls":
-		isPull = optional.Some(true)
-	case "issues":
-		isPull = optional.Some(false)
-	}
-
-	// FIXME: we should be more efficient here
-	createdByID := getUserIDForFilter(ctx, "created_by")
-	if ctx.Written() {
-		return
-	}
-	assignedByID := getUserIDForFilter(ctx, "assigned_by")
-	if ctx.Written() {
-		return
-	}
-	mentionedByID := getUserIDForFilter(ctx, "mentioned_by")
-	if ctx.Written() {
-		return
-	}
-
-	searchOpt := &issue_indexer.SearchOptions{
-		Paginator: &db.ListOptions{
-			Page:     ctx.FormInt("page"),
-			PageSize: convert.ToCorrectPageSize(ctx.FormInt("limit")),
-		},
-		Keyword:   keyword,
-		RepoIDs:   []int64{ctx.Repo.Repository.ID},
-		IsPull:    isPull,
-		IsClosed:  isClosed,
-		ProjectID: projectID,
-		SortBy:    issue_indexer.SortByCreatedDesc,
-	}
-	if since != 0 {
-		searchOpt.UpdatedAfterUnix = optional.Some(since)
-	}
-	if before != 0 {
-		searchOpt.UpdatedBeforeUnix = optional.Some(before)
-	}
-	if len(labelIDs) == 1 && labelIDs[0] == 0 {
-		searchOpt.NoLabelOnly = true
-	} else {
-		for _, labelID := range labelIDs {
-			if labelID > 0 {
-				searchOpt.IncludedLabelIDs = append(searchOpt.IncludedLabelIDs, labelID)
-			} else {
-				searchOpt.ExcludedLabelIDs = append(searchOpt.ExcludedLabelIDs, -labelID)
-			}
-		}
-	}
-
-	if len(mileIDs) == 1 && mileIDs[0] == db.NoConditionID {
-		searchOpt.MilestoneIDs = []int64{0}
-	} else {
-		searchOpt.MilestoneIDs = mileIDs
-	}
-
-	if createdByID > 0 {
-		searchOpt.PosterID = optional.Some(createdByID)
-	}
-	if assignedByID > 0 {
-		searchOpt.AssigneeID = optional.Some(assignedByID)
-	}
-	if mentionedByID > 0 {
-		searchOpt.MentionID = optional.Some(mentionedByID)
-	}
-
-	ids, total, err := issue_indexer.SearchIssues(ctx, searchOpt)
-	if err != nil {
-		ctx.Error(http.StatusInternalServerError, "SearchIssues", err.Error())
-		return
-	}
-	issues, err := issues_model.GetIssuesByIDs(ctx, ids, true)
-	if err != nil {
-		ctx.Error(http.StatusInternalServerError, "FindIssuesByIDs", err.Error())
-		return
-	}
-
-	ctx.SetTotalCountHeader(total)
-	ctx.JSON(http.StatusOK, convert.ToIssueList(ctx, ctx.Doer, issues))
-}
-
-func BatchDeleteIssues(ctx *context.Context) {
-	issues := getActionIssues(ctx)
-	if ctx.Written() {
-		return
-	}
-	for _, issue := range issues {
-		if err := issue_service.DeleteIssue(ctx, ctx.Doer, ctx.Repo.GitRepo, issue); err != nil {
-			ctx.ServerError("DeleteIssue", err)
-			return
-		}
-	}
-	ctx.JSONOK()
-}
-
-// UpdateIssueStatus change issue's status
-func UpdateIssueStatus(ctx *context.Context) {
-	issues := getActionIssues(ctx)
-	if ctx.Written() {
-		return
-	}
-
-	var isClosed bool
-	switch action := ctx.FormString("action"); action {
-	case "open":
-		isClosed = false
-	case "close":
-		isClosed = true
-	default:
-		log.Warn("Unrecognized action: %s", action)
-	}
-
-	if _, err := issues.LoadRepositories(ctx); err != nil {
-		ctx.ServerError("LoadRepositories", err)
-		return
-	}
-	if err := issues.LoadPullRequests(ctx); err != nil {
-		ctx.ServerError("LoadPullRequests", err)
-		return
-	}
-
-	for _, issue := range issues {
-		if issue.IsPull && issue.PullRequest.HasMerged {
-			continue
-		}
-		if issue.IsClosed != isClosed {
-			if err := issue_service.ChangeStatus(ctx, issue, ctx.Doer, "", isClosed); err != nil {
-				if issues_model.IsErrDependenciesLeft(err) {
-					ctx.JSON(http.StatusPreconditionFailed, map[string]any{
-						"error": ctx.Tr("repo.issues.dependency.issue_batch_close_blocked", issue.Index),
-					})
-					return
-				}
-				ctx.ServerError("ChangeStatus", err)
-				return
-			}
-		}
-	}
-	ctx.JSONOK()
-}
-
-// NewComment create a comment for issue
-func NewComment(ctx *context.Context) {
-	form := web.GetForm(ctx).(*forms.CreateCommentForm)
-	issue := GetActionIssue(ctx)
-	if ctx.Written() {
-		return
-	}
-
-	if !ctx.IsSigned || (ctx.Doer.ID != issue.PosterID && !ctx.Repo.CanReadIssuesOrPulls(issue.IsPull)) {
-		if log.IsTrace() {
-			if ctx.IsSigned {
-				issueType := "issues"
-				if issue.IsPull {
-					issueType = "pulls"
-				}
-				log.Trace("Permission Denied: User %-v not the Poster (ID: %d) and cannot read %s in Repo %-v.\n"+
-					"User in Repo has Permissions: %-+v",
-					ctx.Doer,
-					issue.PosterID,
-					issueType,
-					ctx.Repo.Repository,
-					ctx.Repo.Permission)
-			} else {
-				log.Trace("Permission Denied: Not logged in")
-			}
-		}
-
-		ctx.Error(http.StatusForbidden)
-		return
-	}
-
-	if issue.IsLocked && !ctx.Repo.CanWriteIssuesOrPulls(issue.IsPull) && !ctx.Doer.IsAdmin {
-		ctx.JSONError(ctx.Tr("repo.issues.comment_on_locked"))
-		return
-	}
-
-	var attachments []string
-	if setting.Attachment.Enabled {
-		attachments = form.Files
-	}
-
-	if ctx.HasError() {
-		ctx.JSONError(ctx.GetErrMsg())
-		return
-	}
-
-	var comment *issues_model.Comment
-	defer func() {
-		// Check if issue admin/poster changes the status of issue.
-		if (ctx.Repo.CanWriteIssuesOrPulls(issue.IsPull) || (ctx.IsSigned && issue.IsPoster(ctx.Doer.ID))) &&
-			(form.Status == "reopen" || form.Status == "close") &&
-			!(issue.IsPull && issue.PullRequest.HasMerged) {
-			// Duplication and conflict check should apply to reopen pull request.
-			var pr *issues_model.PullRequest
-
-			if form.Status == "reopen" && issue.IsPull {
-				pull := issue.PullRequest
-				var err error
-				pr, err = issues_model.GetUnmergedPullRequest(ctx, pull.HeadRepoID, pull.BaseRepoID, pull.HeadBranch, pull.BaseBranch, pull.Flow)
-				if err != nil {
-					if !issues_model.IsErrPullRequestNotExist(err) {
-						ctx.JSONError(ctx.Tr("repo.issues.dependency.pr_close_blocked"))
-						return
-					}
-				}
-
-				// Regenerate patch and test conflict.
-				if pr == nil {
-					issue.PullRequest.HeadCommitID = ""
-					pull_service.AddToTaskQueue(ctx, issue.PullRequest)
-				}
-
-				// check whether the ref of PR <refs/pulls/pr_index/head> in base repo is consistent with the head commit of head branch in the head repo
-				// get head commit of PR
-				if pull.Flow == issues_model.PullRequestFlowGithub {
-					prHeadRef := pull.GetGitRefName()
-					if err := pull.LoadBaseRepo(ctx); err != nil {
-						ctx.ServerError("Unable to load base repo", err)
-						return
-					}
-					prHeadCommitID, err := git.GetFullCommitID(ctx, pull.BaseRepo.RepoPath(), prHeadRef)
-					if err != nil {
-						ctx.ServerError("Get head commit Id of pr fail", err)
-						return
-					}
-
-					// get head commit of branch in the head repo
-					if err := pull.LoadHeadRepo(ctx); err != nil {
-						ctx.ServerError("Unable to load head repo", err)
-						return
-					}
-					if ok := git.IsBranchExist(ctx, pull.HeadRepo.RepoPath(), pull.BaseBranch); !ok {
-						// todo localize
-						ctx.JSONError("The origin branch is delete, cannot reopen.")
-						return
-					}
-					headBranchRef := pull.GetGitHeadBranchRefName()
-					headBranchCommitID, err := git.GetFullCommitID(ctx, pull.HeadRepo.RepoPath(), headBranchRef)
-					if err != nil {
-						ctx.ServerError("Get head commit Id of head branch fail", err)
-						return
-					}
-
-					err = pull.LoadIssue(ctx)
-					if err != nil {
-						ctx.ServerError("load the issue of pull request error", err)
-						return
-					}
-
-					if prHeadCommitID != headBranchCommitID {
-						// force push to base repo
-						err := git.Push(ctx, pull.HeadRepo.RepoPath(), git.PushOptions{
-							Remote: pull.BaseRepo.RepoPath(),
-							Branch: pull.HeadBranch + ":" + prHeadRef,
-							Force:  true,
-							Env:    repo_module.InternalPushingEnvironment(pull.Issue.Poster, pull.BaseRepo),
-						})
-						if err != nil {
-							ctx.ServerError("force push error", err)
-							return
-						}
-					}
-				}
-			}
-
-			if pr != nil {
-				ctx.Flash.Info(ctx.Tr("repo.pulls.open_unmerged_pull_exists", pr.Index))
-			} else {
-				isClosed := form.Status == "close"
-				if err := issue_service.ChangeStatus(ctx, issue, ctx.Doer, "", isClosed); err != nil {
-					log.Error("ChangeStatus: %v", err)
-
-					if issues_model.IsErrDependenciesLeft(err) {
-						if issue.IsPull {
-							ctx.JSONError(ctx.Tr("repo.issues.dependency.pr_close_blocked"))
-						} else {
-							ctx.JSONError(ctx.Tr("repo.issues.dependency.issue_close_blocked"))
-						}
-						return
-					}
-				} else {
-					if err := stopTimerIfAvailable(ctx, ctx.Doer, issue); err != nil {
-						ctx.ServerError("CreateOrStopIssueStopwatch", err)
-						return
-					}
-
-					log.Trace("Issue [%d] status changed to closed: %v", issue.ID, issue.IsClosed)
-				}
-			}
-		}
-
-		// Redirect to comment hashtag if there is any actual content.
-		typeName := "issues"
-		if issue.IsPull {
-			typeName = "pulls"
-		}
-		if comment != nil {
-			ctx.JSONRedirect(fmt.Sprintf("%s/%s/%d#%s", ctx.Repo.RepoLink, typeName, issue.Index, comment.HashTag()))
-		} else {
-			ctx.JSONRedirect(fmt.Sprintf("%s/%s/%d", ctx.Repo.RepoLink, typeName, issue.Index))
-		}
-	}()
-
-	// Fix #321: Allow empty comments, as long as we have attachments.
-	if len(form.Content) == 0 && len(attachments) == 0 {
-		return
-	}
-
-	comment, err := issue_service.CreateIssueComment(ctx, ctx.Doer, ctx.Repo.Repository, issue, form.Content, attachments)
-	if err != nil {
-		if errors.Is(err, user_model.ErrBlockedUser) {
-			ctx.JSONError(ctx.Tr("repo.issues.comment.blocked_user"))
-		} else {
-			ctx.ServerError("CreateIssueComment", err)
-		}
-		return
-	}
-
-	log.Trace("Comment created: %d/%d/%d", ctx.Repo.Repository.ID, issue.ID, comment.ID)
-}
-
-// UpdateCommentContent change comment of issue's content
-func UpdateCommentContent(ctx *context.Context) {
-	comment, err := issues_model.GetCommentByID(ctx, ctx.PathParamInt64(":id"))
-	if err != nil {
-		ctx.NotFoundOrServerError("GetCommentByID", issues_model.IsErrCommentNotExist, err)
-		return
-	}
-
-	if err := comment.LoadIssue(ctx); err != nil {
-		ctx.NotFoundOrServerError("LoadIssue", issues_model.IsErrIssueNotExist, err)
-		return
-	}
-
-	if comment.Issue.RepoID != ctx.Repo.Repository.ID {
-		ctx.NotFound("CompareRepoID", issues_model.ErrCommentNotExist{})
-		return
-	}
-
-	if !ctx.IsSigned || (ctx.Doer.ID != comment.PosterID && !ctx.Repo.CanWriteIssuesOrPulls(comment.Issue.IsPull)) {
-		ctx.Error(http.StatusForbidden)
-		return
-	}
-
-	if !comment.Type.HasContentSupport() {
-		ctx.Error(http.StatusNoContent)
-		return
-	}
-
-	oldContent := comment.Content
-	newContent := ctx.FormString("content")
-	contentVersion := ctx.FormInt("content_version")
-
-	// allow to save empty content
-	comment.Content = newContent
-	if err = issue_service.UpdateComment(ctx, comment, contentVersion, ctx.Doer, oldContent); err != nil {
-		if errors.Is(err, user_model.ErrBlockedUser) {
-			ctx.JSONError(ctx.Tr("repo.issues.comment.blocked_user"))
-		} else if errors.Is(err, issues_model.ErrCommentAlreadyChanged) {
-			ctx.JSONError(ctx.Tr("repo.comments.edit.already_changed"))
-		} else {
-			ctx.ServerError("UpdateComment", err)
-		}
-		return
-	}
-
-	if err := comment.LoadAttachments(ctx); err != nil {
-		ctx.ServerError("LoadAttachments", err)
-		return
-	}
-
-	// when the update request doesn't intend to update attachments (eg: change checkbox state), ignore attachment updates
-	if !ctx.FormBool("ignore_attachments") {
-		if err := updateAttachments(ctx, comment, ctx.FormStrings("files[]")); err != nil {
-			ctx.ServerError("UpdateAttachments", err)
-			return
-		}
-	}
-
-	var renderedContent template.HTML
-	if comment.Content != "" {
-		renderedContent, err = markdown.RenderString(&markup.RenderContext{
-			Links: markup.Links{
-				Base: ctx.FormString("context"), // FIXME: <- IS THIS SAFE ?
-			},
-			Metas:   ctx.Repo.Repository.ComposeMetas(ctx),
-			GitRepo: ctx.Repo.GitRepo,
-			Repo:    ctx.Repo.Repository,
-			Ctx:     ctx,
-		}, comment.Content)
-		if err != nil {
-			ctx.ServerError("RenderString", err)
-			return
-		}
-	} else {
-		contentEmpty := fmt.Sprintf(`<span class="no-content">%s</span>`, ctx.Tr("repo.issues.no_content"))
-		renderedContent = template.HTML(contentEmpty)
-	}
-
-	ctx.JSON(http.StatusOK, map[string]any{
-		"content":        renderedContent,
-		"contentVersion": comment.ContentVersion,
-		"attachments":    attachmentsHTML(ctx, comment.Attachments, comment.Content),
-	})
-}
-
-// DeleteComment delete comment of issue
-func DeleteComment(ctx *context.Context) {
-	comment, err := issues_model.GetCommentByID(ctx, ctx.PathParamInt64(":id"))
-	if err != nil {
-		ctx.NotFoundOrServerError("GetCommentByID", issues_model.IsErrCommentNotExist, err)
-		return
-	}
-
-	if err := comment.LoadIssue(ctx); err != nil {
-		ctx.NotFoundOrServerError("LoadIssue", issues_model.IsErrIssueNotExist, err)
-		return
-	}
-
-	if comment.Issue.RepoID != ctx.Repo.Repository.ID {
-		ctx.NotFound("CompareRepoID", issues_model.ErrCommentNotExist{})
-		return
-	}
-
-	if !ctx.IsSigned || (ctx.Doer.ID != comment.PosterID && !ctx.Repo.CanWriteIssuesOrPulls(comment.Issue.IsPull)) {
-		ctx.Error(http.StatusForbidden)
-		return
-	} else if !comment.Type.HasContentSupport() {
-		ctx.Error(http.StatusNoContent)
-		return
-	}
-
-	if err = issue_service.DeleteComment(ctx, ctx.Doer, comment); err != nil {
-		ctx.ServerError("DeleteComment", err)
-		return
-	}
-
-	ctx.Status(http.StatusOK)
-}
-
 // ChangeIssueReaction create a reaction for issue
 func ChangeIssueReaction(ctx *context.Context) {
 	form := web.GetForm(ctx).(*forms.ReactionForm)
@@ -3425,112 +2578,6 @@ func ChangeIssueReaction(ctx *context.Context) {
 	})
 }
 
-// ChangeCommentReaction create a reaction for comment
-func ChangeCommentReaction(ctx *context.Context) {
-	form := web.GetForm(ctx).(*forms.ReactionForm)
-	comment, err := issues_model.GetCommentByID(ctx, ctx.PathParamInt64(":id"))
-	if err != nil {
-		ctx.NotFoundOrServerError("GetCommentByID", issues_model.IsErrCommentNotExist, err)
-		return
-	}
-
-	if err := comment.LoadIssue(ctx); err != nil {
-		ctx.NotFoundOrServerError("LoadIssue", issues_model.IsErrIssueNotExist, err)
-		return
-	}
-
-	if comment.Issue.RepoID != ctx.Repo.Repository.ID {
-		ctx.NotFound("CompareRepoID", issues_model.ErrCommentNotExist{})
-		return
-	}
-
-	if !ctx.IsSigned || (ctx.Doer.ID != comment.PosterID && !ctx.Repo.CanReadIssuesOrPulls(comment.Issue.IsPull)) {
-		if log.IsTrace() {
-			if ctx.IsSigned {
-				issueType := "issues"
-				if comment.Issue.IsPull {
-					issueType = "pulls"
-				}
-				log.Trace("Permission Denied: User %-v not the Poster (ID: %d) and cannot read %s in Repo %-v.\n"+
-					"User in Repo has Permissions: %-+v",
-					ctx.Doer,
-					comment.Issue.PosterID,
-					issueType,
-					ctx.Repo.Repository,
-					ctx.Repo.Permission)
-			} else {
-				log.Trace("Permission Denied: Not logged in")
-			}
-		}
-
-		ctx.Error(http.StatusForbidden)
-		return
-	}
-
-	if !comment.Type.HasContentSupport() {
-		ctx.Error(http.StatusNoContent)
-		return
-	}
-
-	switch ctx.PathParam(":action") {
-	case "react":
-		reaction, err := issue_service.CreateCommentReaction(ctx, ctx.Doer, comment, form.Content)
-		if err != nil {
-			if issues_model.IsErrForbiddenIssueReaction(err) || errors.Is(err, user_model.ErrBlockedUser) {
-				ctx.ServerError("ChangeIssueReaction", err)
-				return
-			}
-			log.Info("CreateCommentReaction: %s", err)
-			break
-		}
-		// Reload new reactions
-		comment.Reactions = nil
-		if err = comment.LoadReactions(ctx, ctx.Repo.Repository); err != nil {
-			log.Info("comment.LoadReactions: %s", err)
-			break
-		}
-
-		log.Trace("Reaction for comment created: %d/%d/%d/%d", ctx.Repo.Repository.ID, comment.Issue.ID, comment.ID, reaction.ID)
-	case "unreact":
-		if err := issues_model.DeleteCommentReaction(ctx, ctx.Doer.ID, comment.Issue.ID, comment.ID, form.Content); err != nil {
-			ctx.ServerError("DeleteCommentReaction", err)
-			return
-		}
-
-		// Reload new reactions
-		comment.Reactions = nil
-		if err = comment.LoadReactions(ctx, ctx.Repo.Repository); err != nil {
-			log.Info("comment.LoadReactions: %s", err)
-			break
-		}
-
-		log.Trace("Reaction for comment removed: %d/%d/%d", ctx.Repo.Repository.ID, comment.Issue.ID, comment.ID)
-	default:
-		ctx.NotFound(fmt.Sprintf("Unknown action %s", ctx.PathParam(":action")), nil)
-		return
-	}
-
-	if len(comment.Reactions) == 0 {
-		ctx.JSON(http.StatusOK, map[string]any{
-			"empty": true,
-			"html":  "",
-		})
-		return
-	}
-
-	html, err := ctx.RenderToHTML(tplReactions, map[string]any{
-		"ActionURL": fmt.Sprintf("%s/comments/%d/reactions", ctx.Repo.RepoLink, comment.ID),
-		"Reactions": comment.Reactions.GroupByType(),
-	})
-	if err != nil {
-		ctx.ServerError("ChangeCommentReaction.HTMLString", err)
-		return
-	}
-	ctx.JSON(http.StatusOK, map[string]any{
-		"html": html,
-	})
-}
-
 func addParticipant(poster *user_model.User, participants []*user_model.User) []*user_model.User {
 	for _, part := range participants {
 		if poster.ID == part.ID {
@@ -3574,45 +2621,6 @@ func GetIssueAttachments(ctx *context.Context) {
 	attachments := make([]*api.Attachment, len(issue.Attachments))
 	for i := 0; i < len(issue.Attachments); i++ {
 		attachments[i] = convert.ToAttachment(ctx.Repo.Repository, issue.Attachments[i])
-	}
-	ctx.JSON(http.StatusOK, attachments)
-}
-
-// GetCommentAttachments returns attachments for the comment
-func GetCommentAttachments(ctx *context.Context) {
-	comment, err := issues_model.GetCommentByID(ctx, ctx.PathParamInt64(":id"))
-	if err != nil {
-		ctx.NotFoundOrServerError("GetCommentByID", issues_model.IsErrCommentNotExist, err)
-		return
-	}
-
-	if err := comment.LoadIssue(ctx); err != nil {
-		ctx.NotFoundOrServerError("LoadIssue", issues_model.IsErrIssueNotExist, err)
-		return
-	}
-
-	if comment.Issue.RepoID != ctx.Repo.Repository.ID {
-		ctx.NotFound("CompareRepoID", issues_model.ErrCommentNotExist{})
-		return
-	}
-
-	if !ctx.Repo.Permission.CanReadIssuesOrPulls(comment.Issue.IsPull) {
-		ctx.NotFound("CanReadIssuesOrPulls", issues_model.ErrCommentNotExist{})
-		return
-	}
-
-	if !comment.Type.HasAttachmentSupport() {
-		ctx.ServerError("GetCommentAttachments", fmt.Errorf("comment type %v does not support attachments", comment.Type))
-		return
-	}
-
-	attachments := make([]*api.Attachment, 0)
-	if err := comment.LoadAttachments(ctx); err != nil {
-		ctx.ServerError("LoadAttachments", err)
-		return
-	}
-	for i := 0; i < len(comment.Attachments); i++ {
-		attachments = append(attachments, convert.ToAttachment(ctx.Repo.Repository, comment.Attachments[i]))
 	}
 	ctx.JSON(http.StatusOK, attachments)
 }
